trend_watcher/requirements.txt
google-genai
yt-dlp
python-dotenv
python-json-logger
requests
